LayoutTests/imported/w3c/IndexedDB-private-browsing/idbindex_get6-expected.txt


PASS IDBIndex.get() - throw InvalidStateError when the index is deleted
